Moved to ControllingVideo hook
